Update dashboard.jsp
Cleaned up using declaration tags and java methods. Added some logic to the "ticket's remaining quantity" on the purchase.jsp page, so that it shows "sold out" instead of "0" when there are no tickets left to be purchased. I also hid the "Purchase ticket's" button when all ticket's remaining are 0.